Create a DaraFrame Manager
- To manage all dataframe generation and configuration
