+++ config.yaml
@@ -93,16 +93,16 @@ codes:
   - "SUZB5.SA"
   - "VALE3.SA"
   - "VALE5.SA"
-    #  - "SANB11.SA"
-    #  - "TAEE11.SA"
-    #  - "TBLE3.SA"
-    #- "TIET11.SA"
-    #- "TIMP3.SA"
-    #- "TOTS3.SA"
-    #- "TRPL4.SA"
-    #- "UGPA3.SA"
-    #- "VIVT4.SA"
-    #- "VLID3.SA"
-    #- "VVAR11.SA"
-    #  - "KLBN11.SA"
-    #- "WEGE3.SA"
+  - "SANB11.SA"
+  - "TAEE11.SA"
+  - "TBLE3.SA"
+  - "TIET11.SA"
+  - "TIMP3.SA"
+  - "TOTS3.SA"
+  - "TRPL4.SA"
+  - "UGPA3.SA"
+  - "VIVT4.SA"
+  - "VLID3.SA"
+  - "VVAR11.SA"
+  - "KLBN11.SA"
+  - "WEGE3.SA"
